webnn/pool2dds.html

<!doctype html>
<meta charset=utf-8>
<title>test pooling operations</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-pool2d">
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<script src="./webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  test(() => {
    const x = builder.input('x', {type: 'float32', dimensions: [1, 7, 7, 2]});
    const layout = 'nhwc';
    const y = builder.averagePool2d(x, {
        "windowDimensions": [3, 3],
        "padding": [1, 0, 0, 1],
        "autoPad": "explicit",
        "strides": [2, 2],
        "dilations": [1, 1],
        "layout": "nhwc"
      });
    const graph = builder.build({y});
    const inputs = {
      'x': new Float32Array([
            -68.62557020696815,
            -94.69098419134498,
            -21.020286159397685,
            50.44156641351182,
            15.071063373234892,
            -44.041774749892994,
            -9.447535219701692,
            -55.94677811783235,
            -14.496251290761379,
            84.60435209328762,
            32.08895103610101,
            30.02557486439966,
            12.77174047614156,
            -59.43739359632647,
            -44.74592459902858,
            80.70472647641557,
            9.297827824074247,
            -5.1739910809300795,
            89.91646781270947,
            29.291514962542124,
            -67.17106137197715,
            -9.220385990302475,
            -3.0612675345675484,
            -6.069840480166505,
            79.06002710716876,
            10.560633293664125,
            8.571812725168598,
            -83.74724204696471,
            -68.48414629851965,
            -82.49125080621505,
            -2.1553453974851493,
            -56.933084460286175,
            -2.125115007914772,
            84.70250722493046,
            -39.144590796189235,
            -33.371187301614526,
            -44.54592779846869,
            -40.686313030806566,
            -27.172248050584088,
            -38.596764924548886,
            27.652375471925964,
            -66.86280299047876,
            8.320570333120799,
            -13.53865131680108,
            19.31360929339614,
            85.24203552494563,
            -18.78237401325022,
            -66.89417548435287,
            -16.25868914233442,
            -94.97973255441016,
            -65.36039700301487,
            -80.44886111849867,
            69.06993214251935,
            66.51368330240794,
            92.25735621185515,
            56.692574512549726,
            -34.39699583715945,
            52.732857324738944,
            -14.518666460378597,
            -22.05865850318709,
            -83.52539280923978,
            -9.848713344657114,
            -35.75608243340662,
            27.082336317817862,
            49.987108951102584,
            -77.16894145966208,
            23.242241052087564,
            -51.787454140404996,
            72.20162117050077,
            -90.39583377665323,
            17.914254107150953,
            -16.417600182021275,
            63.21771563714708,
            40.52269549913285,
            -23.723125216545824,
            -23.955614570055687,
            75.96313672248806,
            -14.62187942622188,
            -84.58195628508776,
            -81.89524618494022,
            56.391043657837116,
            -32.7322832219552,
            48.91379183687229,
            48.659523465516315,
            -34.62242302694274,
            -91.44232505260686,
            -84.37861701086531,
            18.940246217480933,
            -50.12198201410851,
            -50.64309604521409,
            -85.15202813282481,
            -38.43618902572565,
            83.70284353724381,
            -90.83333776018745,
            -50.763776934107696,
            -23.699537936690135,
            76.3757883969065,
            98.49473223819575
          ]),
    };
    const outputs = {y: new Float32Array(utils.sizeOfShape([1, 3, 3, 2]))};
    graph.compute(inputs, outputs);
    const expected = [
          -3.351071357727051,
          2.755176067352295,
          1.8019026517868042,
          -0.23048464953899384,
          19.15583610534668,
          -4.010652542114258,
          -1.0493813753128052,
          6.101070880889893,
          -18.503662109375,
          -24.186275482177734,
          15.163517951965332,
          -20.293882369995117,
          -7.353377819061279,
          2.864908218383789,
          -22.448640823364258,
          -46.9700927734375,
          29.124526977539062,
          -26.951427459716797
        ];
    assert_array_approx_equals_ulp('averagePool2d', outputs.y, expected, 11);
  }, 'global averagePool2d nhwc');
</script>
